refactor(libs): made some code clarifications and simplifcations for object cache
Updated and cleaned up some internal functions and expanded the public api slightly. Extended the IBlobCache interface by exported (and implementing a default virtual function) to create (and alloc) cache entry structures. Also exposed a function on CacheEntry to calculate the minimum size of buffer to allocate for entry structures.